+++ SeattleParkingMap/ConstantsExternal.h
@@ -1,0 +1,16 @@
+//
+//  ConstantsExternal.h
+//  Seattle Parking Map
+//
+//  Created by Marc on 1/8/16.
+//  Copyright (c) 2016 Tap Light Software. All rights reserved.
+//
+
+#pragma mark - External Constants
+
+#define SPMExternalAPIArcGISLicenseKey @"ARC_GIS_KEY"
+
+#define SPMExternalAPIBing @"BING_KEY"
+
+//#define SPMExternalAPIHereAppID @"HERE_APP_ID"
+//#define SPMExternalAPIHereAppCode @"HERE_APP_CODE"