+++ SeattleParkingMap/LegendTableViewHeaderFooterView.h
@@ -1,0 +1,11 @@
+//
+//  LegendTableViewHeaderFooterView.h
+//  SeattleParkingMap
+//
+//  Created by Marc on 12/14/19.
+//  Copyright © 2019 Tap Light Software. All rights reserved.
+//
+
+@interface LegendTableViewHeaderFooterView : UITableViewHeaderFooterView
+
+@end