+++ SeattleParkingMapWatch Extension/ParkInterfaceController.m
@@ -78,10 +78,19 @@ static void *ParkInterfaceControllerContext = &ParkInterfaceControllerContext;
 - (void)didAppear
 {
     [super didAppear];
-    
-    [self updateUserActivity:SPMWatchHandoffActivityCurrentScreen
-                    userInfo:@{SPMWatchHandoffUserInfoKeyCurrentScreen : NSStringFromClass(self.class)}
-                  webpageURL:nil];
+
+    if (@available(watchOS 5, *))
+    {
+        NSUserActivity *userActivity = [[NSUserActivity alloc] initWithActivityType:SPMWatchHandoffActivityCurrentScreen];
+        userActivity.userInfo = @{SPMWatchHandoffUserInfoKeyCurrentScreen : NSStringFromClass(self.class)};
+        [self updateUserActivity:userActivity];
+    }
+    else
+    {
+        [self updateUserActivity:SPMWatchHandoffActivityCurrentScreen
+                        userInfo:@{SPMWatchHandoffUserInfoKeyCurrentScreen : NSStringFromClass(self.class)}
+                      webpageURL:nil];
+    }
     
     if (self.needsPickerTimeLimitUpdateOnAppearance)
     {